pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-transformers"
version = "1.0.0"
description = "Transformers para serem utilizados pelas pipelines do projeto."
license = { text = "MIT" }
authors = [
  { name = "Matheus Henrique Dos Santos", email = "matheushenriquep285@gmail.com" }
]
requires-python = ">=3.11"

dependencies = []

[project.optional-dependencies]
airflow = [
    "apache-airflow-providers-docker==4.3.0",
    "apache-airflow-providers-papermill==3.11.3",
    "apache-airflow-providers-postgres==6.4.0"
]

dbt = [
    "dbt-core==1.9.1",
    "dbt-postgres==1.9.1"
]

transformer = [
    "matplotlib==3.10.6",
    "pandas==2.3.2",
    "papermill==2.6.0",
    "psycopg2-binary==2.9.10",
    "pyarrow==17.0.0",
    "pyspark==3.5.3",
    "seaborn==0.13.2",
    "scikit-learn==1.7.2"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["transformer*"]
